backend/requirements.txt
# FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
starlette==0.27.0

# Database and ORM
sqlalchemy==2.0.23
alembic==1.12.1

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-oauth2==1.1.1
authlib==1.2.1
itsdangerous==2.1.2

# Data processing and ML
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
pycaret==3.2.0
joblib==1.3.2

# RAG and embeddings
chromadb==0.4.15
sentence-transformers==2.2.2
langchain==0.0.340
langchain-community==0.0.1
openai==1.3.7

# Async tasks
celery==5.3.4
redis==5.0.1
kombu==5.3.4

# Environment and configuration
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0

# HTTP client
httpx==0.25.2
requests==2.31.0

# Logging and monitoring
loguru==0.7.2

# File handling
python-magic==0.4.27

# Streamlit frontend
streamlit==1.28.2
streamlit-authenticator==0.2.3
plotly==5.17.0
